requirements.txt
PyYAML~=6.0
netmiko~=4.1.2